credential-management/fedcm-network-requests.sub.https.html

<!DOCTYPE html>
<title>Federated Credential Management API basics.</title>
<link rel="help" href="https://wicg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<img src="https://{{host}}:{{ports[https][0]}}/credential-management/support/set_cookie">
<script>
promise_test(t => {
  return new Promise(resolve => {
    document.images[0].addEventListener("error", resolve);
  }).then(event => {
    return navigator.credentials.get({
      federated: {
        providers: [{
          url: 'https://{{host}}:{{ports[https][0]}}/credential-management/support/',
          clientId: '1',
          nonce: '1',
        }]
      }
    });
  }).then(result => {
    return assert_equals(result, "token");
  });
}, "Test that COEP does not break FedCM");
</script>
